proj: Autospec creation for update from version 7.2.1 to version 8.0.0
8.0.0 Release Notes
-------------------

With the release of PROJ 8 the proj_api.h API is finally removed. See
https://proj.org/development/migration.html for more info on how to migrate
from the old to the proj.h API.

With the removal of proj_api.h it has been possible to simplify error codes
and messages given by the software. The error codes are exposed in the API.

Several improvements has been made to the command line utilities as well as
tweaks in the underlying API.

 Updates
 -------

(NEWS truncated at 15 lines)